requirements.txt
streamlit==1.31.1
numpy==1.24.3
opencv-python==4.9.0.80
pillow==10.2.0
tensorflow==2.13.0
pandas==2.0.3
plotly==5.9.0